requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
beautifulsoup4==4.12.2 